Fixed the build - and so much more, read on!
The problem with the build breaking was that Yarn was constantly recreating
orodruin's yarn.lock, probably a bug where file: dependencies using link:
dependencies have unstable (or not properly ignored) checksums.

Anyway, turns out moving the eslint-plugin-node-polyfills above
@rollup/plugin-resolve in the rollup config actually made the issue go away when
using the link: protocol for the orodruin->podigree dependency.

I DON'T KNOW WHY. But I'll commit this as is for now since it should fix the
build. Tests for the issue this caused (so we don't accidentally reintroduce it)
will follow.