Fixed container.js corruption issue (OMFG)
This was a real nasty one! With no error, for no apparent reason, `container.js`
would contain references to `process` (the Node.js built-in), called by a
transitive dependency of podigree.

The relevant change is in rollup.config.js, the rest is babbling. Turns out
rollup-plugin-node-polyfills (which is unmaintained...) hard codes the path to
node_modules, and doesn't follow symlinks within node_modules. Charming... The
current workaround fixes it for now, but for Yarn 2's Plug'n'Play linking, we'd
need another workaround yet.

There's an issue in the rollup-plugin-node-polyfills project for this, but since
it's dead, it probably won't be fixed. Still, the author of that issue probably
saved us hours:

https://github.com/ionic-team/rollup-plugin-node-polyfills/issues/17

Important: There is currently no regression test for this. Presumably the
orodruin/example tests will cover this (once fixed), but for now it can only be
tested manually. Once those tests are fixed, we need to see if reverthing this
fix breaks them, as expected.