Switch to NPM (yup, this is happening)
So, we _did_ switch from NPM to Yarn. That was because of the
container.js issue, where our best workaround was to use Yarn's `file:`
semantics for copying (instead of linking) podigree into orodruin's
node_modules.

Now that this issue is solved, we experimented with some of Yarn 2's
more interesting features, like workspaces and pnp, but ultimately
decided that vanilla NPM is what we want to go for right now. There's a
bit of reasoning in the README now.

So yeah, full circle, but it was worth it.