Don't reference to app bar as nav bar
We probably have to find a better name for the polyPod's navigation bar,
but the navigation bar on Android is the one with the soft buttons on
the bottom of the screen - that's too confusing. The top bar in apps is
generally called app bar, so that's a better name for now and consistent
with the rest of the code.