Implement action callbacks for features
This is not the proper way to do it, it's rather a first version that
works, for now. While the API is in flux anyway, I tried to keep the
number of places in the code base I'm touching for this as low as
possible. In return, it is a little bit hacky.