Apply filters on companyList screen
The filters aren't visible yet on the companyList screen, but the
companies are already filtered accordingly.

This needed some major refactoring - the resulting companyFilter.js is
surely not the most well designed version of itself it could be, but
should do for now. The most ugly part is probably that it extracts
fields (mainly revenueRange) from the company that should probably be
derived properties, rather than having the logic here.