Add average marker and fix gauge curve
Would have been nicer to do this in separate commits, but not really
worth the effort.

The fill of the gauge used to have no border radius, that looked weird,
so now the left side is always round, the right side becomes round at
>98% fill (that's the value at which it looked good for me).

The CSS will need some refactoring.