Support API level 24
I've tested various Android versions, and without any changes,
polyExplorer currently only works on API level 28, which supports only a
fraction of the devices out there.

In the end, I decided to include a polyfill for Object.fromEntries in
the polyPod itself, to keep the burden of supporting old WebView
versions low on feature developers. If we find more stuff that's
missing, we can add more polyfills. We should however consider to move
away from the system WebView - as long as we use it, features need to be
tested very carefully on older Android versions.