Create a Screen component
We had a bit of duplication between the screens, which got a bit crazy
when we integrated that shadow. The new screen and topShadow components
reduce duplication a bit.

I did _not_ make these changes for the companyInfo screen to avoid merge
conflicts when changing the indentation level, but I did add a TODO.