Refactor FeatureStorage
Still not brilliant, but at least there is now only one Feature (data)
class, and feature loading is in one place.

There is still no state anywhere to be seen - in theory we could just
have one instance of FeatureStorage and use it from both fragments that
need it, then we wouldn't have to load features twice. But this seems
good enough for now.