Calculate distance from label to correlation center correctly
Previously, the distance from label center to correlation center was
considered, now we are considering the distance from the label's
bounding rect to the correlation center.

Labels can _still_ clash, this either needs collision detection, or a
move away from the top/right/bottom/left position snapping. We can leave
it alone until we find a case where it's broken.