Simplify description/category merging
I created patch-data.js, which creates company data in the expected
format - all the convert script has to do is merge entities it finds
there with the entities parsed from polyPedia.

While it's "simpler", it is not necessarily "easier": There's a certain
appeal to a dedicated list of descriptions and categories. But we can
still reformat the patch-data.js file to generate the required format
from separate objects if we want to.