Fix PROD4POD-320: Enable tap to swipe for all non-interactive slides
I created an actual DOM element for the tap area, so it's easier to
debug. The general idea was to make the tappable area a bit smaller than
the slide, to avoid accidentally going to the next slide when trying to
hit the progress bar or the 'how to read this' sections.

I disabled this completely for the company list and purpose slides,
since there's a lot to tap on those.