Prepare to show industry categories on multiple slides
This doesn't actually change the behaviour yet, just increases the data
exploration loading time a little bit really. But I wanted to already
test that out in practice, and not land this as such a huge change.

There's various TODOs in companyBubbles.jsx now - with these solved
we'll have the desired behaviour without having to touch anything else.