Fix polyEpoch calculation in Android build
Of course Java's odd Date classes had to bite me again, after all those
years of avoiding them :D The problem was that the founding date was at
midnight in the local time zone (CEST in my case), not UTC.

That seems to be the nicest way to do it in Groovy. I did notice that
when converting this code to Kotlin, it behaves as expected (by me).