Make whitespace and indentation consistent
I didn't touch MessagePack with any of this, since that appears to be
third party code.

How empty lines were inserted was a bit inconsistent in the original
code base, I manually corrected that.

For indentation, I simply used XCode's auto indentation, which
unfortunately produces trailing whitespace for empty lines in
blocks. Not too elegant, but consistency is more important.