Re-apply 5e8ff and 7737a - fixed some typos while at it
This is the last step of resolving the issue where we had accidentally
pushed a new version of companies.json to main (250f9c3) and then
accidentally merged it into release (b2abc01).

5e8ff and its followup, 7737a, were some manual fixes we had made to
companies.json in the release branch - not super elegant, but a
conscious decision.

Those commits didn't cleanly re-apply on top of the reverted
companies.json, so I made the changes again manually. I couldn't resist
fixing a few typos in those changes in the process :)