Fix: Status bar foreground is light on iOS 14
Turns out the code we have for parsing colours in iOS 13 didn't work in
iOS 14, since there it appears a colour with the hex value #FFFFFF is
automatically transformed to the named color "white" - which breaks our
colour component parsing.

To solve this, we are now falling back to the native implementation of
getting colour compoments on iOS 14 and later, while sticking to the
hacky approach in iOS 13.