Move podjs to the top level
So far, none of the code under core is shipped to end users directly -
whereas pod.js would be (if only to developers).

Secondly, pod.js is the first version of the web-based pod - which is
one of our platforms next to Android and iOS, which are all on the top
level so far.