Implement polyNav API in BrowserPod
For starters, the implementation has largely the same logic we had in
polyExplorer's fakePod.js.

I originally added the actions property to the PolyNav interface while
at it, and while it worked, tracing.ts in podigree just wouldn't compile
anymore, so I gave up on that, given how we'll most likely redesign this
API soon anyway.