Introduce CompanyFilter class
This is the first step of the refactoring I wanted to do: Moving away
from all these functions that modify a map of filters and have to
account for inconsistent data from the caller in return. This approach
allowed me to get rid of ensureSet(), which was pretty nasty.

I also added a test for sortedMap (previously sortFilters) while at it.

I didn't resolve two existing issues I noticed yet, but added TODOs:

1. Industry category names are built while companies are being
   extracted, in global state. Seems like a lookup from globalData is
   how it should really be.

2. There's a company in the data whose countryCode is `false` - I kept
   the workaround for now.

As for what I still want to refactor: The code in fields is still pretty
nasty, and I'm thinking about whether to introduce some additional
objects to interact with, rather than most API calls requiring a field
and a value.