Introduce CompanyFilter class
This was a refactoring I had been dreaming about for a while: Previously
we just had a bunch of functions that accepted a filters object (which
could either contain values as sets or arrays...). Now that's all
encapsulated away, but the external interface is still largely the way
it was before.

While at it, I increased test coverage a little bit (sorting was barely
covered), and added a few TODO comments for things that seemed error
prone to me.