Extract remote-pod from podigree
podigree contained a soup of Pod implementations and produced the
bootstrap.js file, which is used on both Android and iOS to set up
communication between Feature and Pod. It's so central, it deserves its
own package.

The remaining Pod implementations in podigree are not being used
anywhere, but we'll leave them around for inspiration for now.