Move fetch-spec, rdf*, customs and poly-api core core/api
While at it, I changed customs and poly-api to their new names:
manifest-parser and pod-api respectively.

This was a slightly larger operation, I had to regenerate pretty much
all package lock files, and in order to do that for fetch-spec, I even
had to downgrade the jest version used there - it was in conflict with
the ts-jest version, and I didn't feel like adding that fix to this
already pretty bulky commit that ultimately just moves a few directories
around.