Fixing compiling warnings that (maybe) led to errors in remote-pod
Essentially my rationale here is that msgpack, that was excluded as a
"peer import" in esm and commonjs format, wasn't in 'iife' format, so
maybe it should be treated in the same way.

That eliminated the warning, and it still passes the
tests. Essentially because, as I understand, msgpack now is *not*
bundled (because `external`), so no way it produces a warning anywhere.

However, I'm treading on thin ice here because:

1. I have no idea what `iife` does, or why three different files are
generated.
2. There are no integration tests for the generated file. At this
level, or maybe at a higher level, they should be tested.