Fix typo
Well, the larger story is that we're being a bit inconsistent about
capitalising _Feature_ in these READMEs. In most places, we don't
capitalise _feature_, and we refer to _the polyPod_ instead of _the
Pod_. That's different from the technical whitepaper, but that seems
alright. What's not alright is that it's inconsistent within our
READMEs. But as long as we're not inconsistent in one file (like here),
that seems acceptable for now.