:recycle: to create custom exception :boom:
If the hash didn't have the correct language key, it was not going to
have a key with the default language, so the second sentence in the
constructor was going to simply fail. I still need to run some
integration tests with this, because it might be a hack to make
something else work, but it definitely looked weird.

Anyway, this adds a custom exception, which is probably better in a
big framework, to not have `Error`s cropping up somewhere else to deal
with them based on message, while at the same time keeping 100%
coverage.

This is (possibly) a breaking change, because some code could rely on the old
behavior. Let's see.