Separate all Android/iOS licenses by two empty lines
Same logic as we have in js-licenses.txt. The only differences is that
js-licenses.txt comes with two empty lines at the end as well, something
most editors remove, but since the file is generated I didn't want to
edit it at this point. We can fix that in the script that generates the
file.