Try a new staged build
That will trigger build of features only when downstream dependencies
have changed, also checks for an array of versions. Eventually, we
will have to create a web of dependencies that trigger only when
needed.

Also, jobs for polyExplorer and lexicon are run in parallel, so it's
not taking the whole time to test them.