Read feature display order from bundle
Still a bit hacky, but better than having to hard code the order in
multiple places.

While at it, I moved all files from the feature bundle to a "features"
subdirectory on iOS - now that we have a file not ending in .zip in the
bundle, it gets too messy having everything at the bundle's top level.