Add warning in action name
Main problem I see with this is that it will fail *always* once you've tried to
change one of the frozen files, so the only way around is to rebase and
force-push or simply close the PR and start all over again (possibly after
squashing changes).

I don't see any easy workaround for this, or even that we would want a milder
failure, because they could be more complicated (for instance, looking at
actuall diffs between HEAD and the last commit) and might have some workaround
we might not think of. This approach seems secure, if a bit harsh.