Uses jsdoc for documentation :memo:
This is the story of a failure. I've tried for a good while to make typedoc
work. There's no way it works with plain js, apparently. Before I was using
tsdoc, which is also for typescript, but actually uses js underneath. I wanted
to change that because jsdoc has some vulnerabilities, while typedoc has not.

Eventually, what the heck, jsdoc is standard, lots of plugins, it's mature and
well, it's just the tool for the job. So there's that.