Moves the root file
Which was apparently used only to import a single file from the test files. This
is not considered a best practice, which is why only the concerned directory has
been imported. As a matter of fact, not ideal either, since it would be best to
import only the modules that are under test.