Deleted fonts used only in a storybook "story"
Again, drawing attention to PROD4POD-525. These fonts were included in the repo
only for something that does not have a clear function in the repo, that's not
tested, and, as far as I understand it, these specific fonts don't play any
specific role in the story; any other font could possibly be used, since it does
not affect the functionality under (test|prototyping).