Fix remote-pod build
Originally, pickFile() had a return type of void, as opposed to
`Uint8Array | null`. While I had already fixed that in pod-api, I forgot
to also address it in remote-pod.

Now, passing a union type for `ValueEndpointSpec` turned out to be quite
the adventure. I had to fix an issue with `ForcePromise` in postoffice,
since it distributed union values. For example, the following type:

    ForcePromise<Uint8Array | null>

Would expand to:

    Promise<Uint8Array> | Promise<null>

as opposed to:

    Promise<Uint8Array | null>