fix: remove babel dependency
The idea of using babel js to allow the use of private attributes and methods has resulted in a big mistake.
Babel has a strong dependency on the core-js and regenerator-runtime libraries that force us to add them
within poly-look or/ and in the features. Although we have tried to add these dependencies in the library and in
the features, we could not solve all the dependency problems generated by babel js, so we decided to permanently
remove it from the poly-look dependency and dispense of the private attributes and methods