Less :coffin:
This was actually from a spike to automatically load locales from the FS. It didn't work on the browser (quite obviously, although with our "fake" fs it is not totally lost for the future), but apparently I forgot to eliminate the test and the data used in the test. All test data is actually generated on the fly in the remaining test, so this also addresses one of @fhd concerns.