Adds general/common structure
I have substituted directory names that seem to follow a certain regex (included
in this commit) by uniqueid_hash. This uniqueid is either an user name, or
facebookuser<number>, or a combination of names in case of group
conversation. It can also be the name of an album. This includes personal
information, so it's better left, but fortunately always follows the same
pattern.
Some uniqueis that don't follow that pattern are still false positives. I
haven't seen any false negative, at least in the data I have.
It still might need another iteration to avoid those false positives, I guess.