Better file structure schema
Still, I don't like the generic names because they can't be matched exactly to
what's on the file. For the time being, they allow us to know the structure we
can find. We can also generate test files from this.

Possibly, the best way to carry this out would be a JSON schema. I'll try and
find a way to generate one.