Can't really use zip.js with node
So we'll go for second best, and only with test purposes. Anyway, this still
does not work. I need to find out why file has a "time" attribute, and where I
can create that kind of object. Gut it's a checkpoint to find out a bit more,
and where everything runs together nicely.