Fix Android pod initialisation race condition
It appears that on Android, in some situations (at least the Pixel 2
emulator running Android 30), it can happen that the RemoteClientPod is
already initialised and starts to send messages through the browser
contexts, but the other side (i.e. in Android) is not ready to receive
messages yet.

Not a particularly beautiful fix, but it's better than having to add
random initialisation sleep calls in features.