PROD4POD-726 - Return success value from network.httpPost
In facebookImport, I left it at a simple console.log with the success
value for now.

On Android, while signaling an unsuccessful request correctly, there are
two issues:

1. httpPost is broken on Android in general, so it can never succeed. I
wanted to tackle this separately

2. The logic for waiting for the response status code and checking it
for success is missing. Again - wasn't something I could really
implement when we never get a response