[Task] Adds script for basic generation of a zip test file (#242)
* Adds script for basic generation of a zip test file

This only generates *the structure*, based on the common folders found. It does
not create any kind of reasonable content. However, it's a start and it does
work with the files that summarize the Facebook file structure

(the change to .gitignore in another folder just slipped by)

* Adds script

* Fixes test for emptyness

* Essentially, this moves a constant from the old "globals"

- A new directory has been created instead of a file to preserve the current
folder structure.
- There was some refactoring involved to have it used from a script, and this
needed to declare the current feature as a `module`. As far as I have seen,
`polyExplorer` is also declared as a module, and this will help later on when we
start to create unit tests.
- This new constant has been used where one literal was involved.

* Removes this file after the last commit

The only thing that was actually used has moved to where it belongs :truck:

* Deletes superseded files

* :truck: to canon directory

* Adapts script to new globals :recycle: