PROD4POD-787 - Add error handling for importing/reading
I left the actual importers and analyses alone, since their error
handling feeds directly into the report. Since we don't have named error
types coming from the backend, we'd still have to include the message in
those to really be able to debug problems, however.

The global error popup is a bit ugly at this point, but I guess it's a
start.