Revising y eliminating :coffin:
Mocha was not really invoked in any script; there was no `test` target in npm.
There's a single file, which includes a type, `Java.type`, which apparently would only work if invoked from  Java implementation of JavaScript.

So unused dependencies have been eliminated, as well as unused code. Since fetch is deprecated, the whole module might also be dead code, but that's outside the scope of this PR