This was tricky
the `lint` target was not only installing lint, but also running ci on the root package.json. Originally it only included linting, but it now includes jest... which was not installed and was thus making the previous tests fail.

I have moved the installation to outside, which probably makes sense, although it would probably make much more sense if every target would include checking if node_modules exists or not.