WIP: Implement HTTP based readFile (#338)
Refactor the way features interact with the file system by introducing HTTP-based interface. This improves the preformance with which the data flows into the feature.

--- 
* Implementat HTTP based readFile

* iOS implementation

* Refactor

* Address comments

* Fix entry listing on overview screen

* Fix the test

* Actually fix the test

* Fix: Error loading previous import

We changed the protocol from polyPod: to polypod: while introducing the
new HTTP-based readFile, this change makes it work for users who already
had an import.

* Fix: Empty file from previous import

On Android, we moved away from polypod: URLs while introducing the new
HTTP-based readFile. This change makes it work for users who already had
an import.

Co-authored-by: Felix Dahlke <felix.dahlke@polypoly.net>