Add --start option to build.js
This will allow us to do a manual partial rebuild when working on core
components quite a bit easier.

I also couldn't resist to finally introduce some tests for the build
logic. They are arguably not incredibly elegant, written against the
current state in the repository rather than test data. But I'd argue
it's a good start as any.