[Task] [PROD4POD-507] Trying to eliminate/unify binary assets from the repository. (#133)
* Deleted fonts used only in a storybook "story"

Again, drawing attention to PROD4POD-525. These fonts were included in the repo
only for something that does not have a clear function in the repo, that's not
tested, and, as far as I understand it, these specific fonts don't play any
specific role in the story; any other font could possibly be used, since it does
not affect the functionality under (test|prototyping).

* Actually, this one is not even used :coffin:

* Remove used font, substituting by copy using rollup

* :see_no_evil:

* Change config to copy from android directory, including :see_no_evil:

* More unification of assets

* Moving also Android assets to another location :truck:

* :recycle: :truck: for new location

* Adds another dependency :bug:

* Explain why we don't want fonts in the .gitignore commit

And also add OpenType fonts; this only included TrueType fonts.

* Changing the name of the assets dir to, surprise, assets

* Reverting gradle bump

* :bug: path

* So you had to go and save them...

* Add exception in fonts directory

* Adding as symbolic links

* Restores original build

* Restores original slide

As well as original fonts. Copying them to the place was a code smell, so I'll
leave like this until I figure out what to do with them.

* Testing with linked directory

* Use /assets/fonts in iOS

* Eliminate polyDisplay font

It has been removed from the main branch, but this branch would have
accidentally reintroduced it.

* Changed to symlink

* Eliminates copying, uses directly symlinks

* Eliminates :warning: following advice

* Changes another rollup

* :coffin: ignore pattern

* :coffin: ignore pattern

Co-authored-by: Felix Dahlke <felix.dahlke@polypoly.net>