Fix UI tests
Now that the notification data is being read from an asset file, the UI
tests (which run in a separate app) cannot access the same content any
longer. As a workaround we're just looking for the close button now - if
there's another button with that label on the screen, the tests that
check that no notification is showing up would fail, so seems robust
enough.