PROD4POD-386 - Add push notifications on iOS
I'm not actually entirely sure if this works.

There is a way to test background tasks by triggering them manually in the Xcode
debugger - there everything works fine. But I haven't been able to reproduce a
situation where the background task runs on its own. iOS decides if and when to
run background tasks on a variety of factors, from battery level to how
frequently the app has been used.

Since I can't seem to get the background task to run on its own, I also wasn't
able to test that it runs _after_ an update, but _before_ the app was launched,
which we need for this logic to work.

Landing the logic for now, so that we can test it with some more people.