PROD4POD-913 - Add thumbnailColor to feature manifests
Feature thumbnails are always rendered in the correct aspect ratio -
rather than cut off. As a result, we have some background space to
account for to fill the feature's card.

For this, we used to just use the feature's primaryColor, assuming that
the thumbnail's background colour would always match that one. Turns out
that's not the case, so I gave up and added a dedicated thumbnailColor
field that can be used to overwrite this. If not specified, we fall back
to primaryColor, i.e. the old behaviour.

I specified it for the two features that had a mismatch: facebookImport
and polyPreview.