PROD4POD-1057/1061 - Change security flow on Android
While at it, I couldn't quite resist eliminating some duplicated logic
and unused code, as well as moving the authentication related code to a
dedicated class - similar to how it was implemented on iOS.

Additionally, I noticed some weirdness with the onboarding activity
deciding whether it should close itself, it would still quickly light up
in my tests. So instead I changed the logic for whether the onboarding
is shown a bit.