PROD4POD-1069 - Only show security onboarding after first run
While working on this, I also found some unused code that checks whether
a pass code is available on the device. Added it as an additional check,
but the dedicated check to see if the app is running in the simulator is
still needed. Doesn't seem to hurt though.

I also _really_ didn't want to refactor how we deal with user defaults
at this point, but I got sufficiently fed up that I added a TODO
comment.