Eliminate empty app state on iOS
As previously implemented, if the user would _cancel_ the
authentication, they would be left with a completely empty app. They
would have to force quit it and reopen it to use it.

This is a workaround until we have time to implement a better solution,
probably a dedicated screen for the locked state.