PROD4POD-689 - Update privacy policy
While it's not being fetched automatically yet, we're very close now:
The plain HTML version of the privacy policy is being generated on
polypoly.coop, and I've moved the actual content to /assets, to be used
by both Android and iOS.

All that's missing now is essentially a script that updates the
respective files in /assets.

While working on this, I also extracted the language determination
logic, which was missing fallback logic on iOS.

Theoretically, we could also move the other legal texts to /assets now,
but I figured it's already a large enough change.