PROD4POD-1131 - Reorder and clean up feature list
Changed things to the desired order, and also hid polyExplorer V2 and
demo while at it.

On Android, we would still show features present on disk but not present
in the generated order file. Changed this behaviour for now, since we
got into the habit of adding features temporarily for testing purposes.