PROD4POD-1097 - Relentlessly authenticate on Android as well
Originally it seemed like a bad workaround on iOS to just authenticate
in a loop, but turns out in practice it's a lot better than just closing
the app after one failed attempt.

Turns out the solution was pretty simple - just relaunch the activity
when authentication fails. I'm mildly surprised it works as intended,
but according to my tests it does.