Add proper logging on iOS
This should make it easier for us to understand crashes we cannot
reproduce. In my understanding, all the error level log messages we
write, are being stored on the device. Info level log messages aren't,
but I think all the cases I used info for are not incredibly helpful for
debugging issues that happen in the wild.

I did improve _some_ log messages, but there's still a lot of work to do
in terms of making them consistent, not duplicating the same strings so
much, and being more specific about where an issue occurred. A large
chunk of our error log messages in the API code also needs to throw an
actual error, so that it can propagate back to the feature. As opposed
to Android, a lot of errors are currently being swallowed.

But still, I think this is a good first step.