Print per-package build timings (#526)
This should help us understand which parts of the process are slow.

While at it, I also reduced the noise a bit by adding
--no-update-notifier (and --no-fund, for good measure), to all NPM
invocations.

There's room for improvement though:

1. I experimented with adding a --quiet option for build.js that
surpresses all of the actual build output, but took longer than a few
minutes so I didn't bother at this point. It's easy enough to comment
out the console.log statements in executeProcess temporarily :P

2. We're invoking NPX for linting - that doesn't print any timings ATM,
but I suppose we're not too interested in those execution times so I
didn't bother for now.