Fix Android tests
Some old Android integration tests currently don't compile - they were a
bit experimental in the first place, and never got updated along with
other refactoring we did, because we never ran them.

Recently, we started to add some more meaningful integration tests we
actually _do_ want to make sure we run. So in order for it to make it
easy to run these, I've commented out the broken ones for now. Only a
temporary measure for sure.