Fix feature bundling
While I don't test if source "." works on Windows (to my knowledge it
should), it certainly seems to be required on POSIX systems.

Also returned the promise we get from "zip" for good measure - we're
ignoring it and it's actually safe to ignore, we can bundle all the
features and write the "order" file in parallel, as it happens with this
change. But still :)