Move duplicate behaviour to a core package/group
First of all: Naming is hard, definitely for me. We already have /core
in the repository, and while it is _meant_ to hold shared behaviour in
general, in practice, it doesn't. It solely holds code executed in
features, no code executed on the platform level.

For shared platform behaviour, we aim to find a good approach of using
the exact same code on both Android and iOS. In preparation of that
step, we're already extracting shared, platform independent logic into
Swift/Kotlin components whose interface and tests (and usually
implementation) are essentially identical.

The only true example of that is currently the update notification
logic, but with a little bit of work, the feature manifest parsing logic
can soon follow. And we should be able to find more shared behaviour to
tease out in other parts of the code base, over time.

Since there's more shared behaviour, more "similarly" implemented logic
that we want to ultimately get rid of, I created a core package
(Android) and group (iOS) to hold all of that code, so it's easy to
identify.

But yeah, naming is hard.