Returned to platform-dependent code
Only I've changed it to affect all possible commands in executeProcess. This was causing a bootstrap problem: you needed to run an npm ci to run an npm ci. And at any rate, we were going from no-deps build.js to a single dep build.js. So it was either this or having to run npm ci before running this. Possibly not a bad deal, but also a phase shift I didn't feel comfortable with. So between two uncomfortabilities, I chose the one closest to the Windows, which is this one, I guess. Sorry for all the trouble in between...