Prod4 pod 1177 (#547)
* changed the npm call when a task is being spawned by npm to fit windows environment

* symbolic links changed to accomodate windows

* latest windows build fixes

* fonts showing the right way round on the onboarding slide

* fixing build on windows using npm packages. Introduced dependencies 'shx' to run linux shell commands and 'bestzip' to zip files. Dependencies help maintain the code clean without a lot of workarounds involving the npm scripts

* reverted symlink removal

* zipping functionality changed from an execFileSync to actual use of the bestzip package

* removal of windowsEnvironment variable

* Fix clean script (requires a root module now)

* Eliminate unrelated changes

* Fix symlinks

* Fix feature bundling

While I don't test if source "." works on Windows (to my knowledge it
should), it certainly seems to be required on POSIX systems.

Also returned the promise we get from "zip" for good measure - we're
ignoring it and it's actually safe to ignore, we can bundle all the
features and write the "order" file in parallel, as it happens with this
change. But still :)

* Clean root package last

* shelljs call replacing childProcess.spawn as to avoid workaround with identifying windows evironment

* Returned to platform-dependent code

Only I've changed it to affect all possible commands in executeProcess. This was causing a bootstrap problem: you needed to run an npm ci to run an npm ci. And at any rate, we were going from no-deps build.js to a single dep build.js. So it was either this or having to run npm ci before running this. Possibly not a bad deal, but also a phase shift I didn't feel comfortable with. So between two uncomfortabilities, I chose the one closest to the Windows, which is this one, I guess. Sorry for all the trouble in between...

* Ensure root package actually gets cleaned

* Don't waste so much space

* Returning color to our _lives_builds

Co-authored-by: unknown <alex.farhat.95@gmail.com>
Co-authored-by: Felix Dahlke <felix.dahlke@polypoly.net>
Co-authored-by: Juan Julián (JJ) Merelo <juan.guervos@polypoly.net>