Use EntityList in CompanyIndustryList
For now at least, I had to leave the fallback in place - the infinite
scrolling logic isn't very efficient when one group has a large amount
of elements, but I figure it's worth it anyway.

To accommodate the new variants we need in EntityList to support the
CompanyIndustryList use case, I added an option for moving the group
label to the side - the default behaviour being the group label on top.

Since it felt wrong to hard code that, I also added a function that is
responsible for creating the component, preprocessing the input and
setting some options (if not set by the caller) based on that. I think
this approach - having things be configurable, but also having flexible
defaults, could be interesting.