Eliminate DefaultPod
It's not actually being used anywhere, and we constantly need to update
it when we change the API.

A default implementation doesn't make a lot of sense for the polyPod
API, given how every implementation is highly platform dependent, and
given how the actual interface is not some JS code, but the API exposed
to features.

True spec tests would need to be written as a feature, to be able to
exercise all polyPod implementations for validity, but for now, I kept
the existing spec tests and am just running them against BrowserPod -
which is kind of our reference (not default) implementation in practice.