Fix: demo feature not cleaned properly
The demo feature was exposing a "clean" script, which did however not do
a full cleaning, it was just being used as a prebuild step, specifically
to remove the dist directory. Our build system, however, took the
existance of a "clean" script to not apply its default logic of removing
"dist" and "node_modules". Rectified that.