PROD4POD-664 - Finish Companies sections (#569)
I had to add some more features to BubbleCluster, primarily:

1. Bubble nesting (was _almost_ already supported)
2. Setting the stroke of bubbles
3. Setting arbitrary text for bubbles

This required two changes in the facebookImport feature, since it was
using an attribute that is now gone. The default behaviour, however, was
left as it was.

For the bar chart, I tried to use vertical labels, but it seems non-trivial,
so I ended up using the indexed legend there as well, seems acceptable.