Improve API error messages
Was a bit tricky to do this without rewriting this whole switch/case
jungle we have here. While the resulting messages aren't the very most
user friendly with this approach, at least they allow us to better
understand what's going wrong when an API call fails, and we don't rely
on something as error prone as including the name of each API directly
in the error message.

I didn't really touch logging - if we centralise error handling in the
future, we can also centralise logging.