Add version check in build.js (#598)
* Moves to own function and add emoji

* Adds checking of npm version

* Adding some debug code

* Eliminate debug

* Eliminate :construction:

* Don't exit from subs

* Still :bug:

* Converting to string

* :fire: going

* :coffin:

* Better practices returning exitCode

* Address positioning of manifest reading

Does not seem to have any influence...

* Give npm a try

* add platformize to DRY command processing

* Brings semver to the forefront

It was already installed in many packages, so this is not adding a new one

* Auto audit fix

* Changes version, uses semver for fun and profit

* Uninstalling, although this was not really necessary

* Revert using semver

And also revert to using just major versions. We can't use semver, since build.js needs to use only core stuff; semver broke lint, for instance, since we only require npm ci in some commands. We could add lint there, but not worth the while. Also, we could semverize (can still do), but at the end of the day these major minimal version is all we're going to need