P4P-1278: fix broken font-weight
somehow the font-weight of Jost Medium and Jost Bold behaves like a font-weight: 600 although I could not find the source for it, there is no font-weight 600 anywhere in the code, from where it would affect these cluster-story spots. Nonetheless putting font-weight: 400 in the regarding spots, solves the problem for now.