PROD4POD-1380 - Fix/stop barchart redraw
The chart was rendered twice when returning from the entity-details page. Since
the data given to the chart remained the same, the update function was fired up
on the second render (rather than the entering function that was called in the
first render). The update function was broken and would break the chart on the
second render. Fix was to explicitly select barValues when updating them.