Change conditions for release-*
Now (and before) they were repeated, and there's no easy workaround for that, see here https://github.com/actions/runner/issues/662. Early exit would be desired, but the only alternative is to define an environment variable, which would be another step and would still need to check the value; or fail the action, which wouldn't be a good thing to see. So, there's no way to DRY here apparently